Enforce strict architecture checks and split provider tests
