+++ .github/workflows/ci.yml
@@ -30,5 +30,11 @@ jobs:
       - name: Test
         run: cargo test --workspace --all-features --locked
 
+      - name: Dependency graph contract
+        run: ./scripts/check_dep_graph.sh
+
+      - name: Strict architecture contract
+        run: LOONGCLAW_ARCH_STRICT=true ./scripts/check_architecture_boundaries.sh
+
       - name: Docs build
         run: cargo doc --workspace --no-deps
